add controller to Balance and Imbibition

+++ database/migrations/2022_08_29_133330_create_balances_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('sugar_cane')->nullable();
             $table->integer('totalizer_raw_juice')->nullable();
             $table->integer('flow_raw_juice')->nullable();
-            $table->integer('raw_juice_percent_sugar_cane')->nullable();
+            $table->float('raw_juice_percent_sugar_cane')->nullable();
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
         });
